requirements.txt
pydantic>=2.8.2
openai>=1.40.0
anthropic>=0.34.0
rapidfuzz>=3.9.0
rich>=13.7.1
typer>=0.12.3
pandas>=2.2.2
python-dotenv>=1.0.1
httpx>=0.27.0
pytest>=8.3.2
